+++ src/luw-mapping.ts
@@ -1,22 +1,15 @@
-import { BigInt, Bytes, log, store } from '@graphprotocol/graph-ts'
+import { BigInt, Bytes, log } from '@graphprotocol/graph-ts'
 
 import {
   AddedOwner as AddedOwnerEvent,
   ChangedThreshold as ChangedThresholdEvent,
   CreateLicensedUserWallet as CreateLUWEvent,
-  ExecutionSuccess as ExecutionSuccessEvent,
   MigrateLicensedUserWallet as MigrateLicensedUserWalletEvent,
   RemovedOwner as RemovedOwnerEvent,
-  RequestTypeRegistered as RequestTypeRegisteredEvent,
-  RequestTypeUnregistered as RequestTypeUnregisteredEvent,
   UpdateLicensedUserDetail as UpdateLicensedUserDetailEvent,
   UpgradeLicensedUserWalletToMulti as UpgradeLicensedUserWalletToMultiEvent,
 } from '../generated/LicensedUserManager/LicensedUserManager'
-import {
-  LicensedUserWallet,
-  MetaTxExecution,
-  MetaTxRequestType,
-} from '../generated/schema'
+import { LicensedUserWallet } from '../generated/schema'
 import { eventUTCMillis, logInvocation, secondsToMillis } from './utils'
 
 function userType(n: i32): string {
@@ -184,57 +177,6 @@ export function handleUpdateLicensedUserDetail(
 
   luw.updatedAt = eventUTCMillis(event);
   luw.save();
-}
-
-export function handleRequestTypeRegistered(
-  event: RequestTypeRegisteredEvent
-): void {
-  logInvocation("handleRequestTypeRegistered", event);
-
-  let typeHash = event.params.typeHash;
-  let requestType = new MetaTxRequestType(typeHash.toHexString());
-  requestType.typeHash = typeHash;
-  requestType.typeString = event.params.typeStr;
-
-  log.info("creating MetaTxRequestType hash [{}] typeString [{}]", [
-    requestType.typeHash.toHexString(),
-    requestType.typeString,
-  ]);
-  requestType.createdAt = eventUTCMillis(event);
-  requestType.save();
-}
-
-export function handleRequestTypeUnregistered(
-  event: RequestTypeUnregisteredEvent
-): void {
-  logInvocation("handleRequestTypeUnregistered", event);
-
-  let requestTypeId = event.params.typeHash.toHexString();
-  let requestType = MetaTxRequestType.load(requestTypeId);
-  if (requestType == null) {
-    log.error(
-      "received ReqestTypeUnregistered for unknown MetaTxRequestType: {}",
-      [requestTypeId]
-    );
-    return;
-  }
-
-  log.info("removing MetaTxRequestType hash [{}] typeString [{}]", [
-    requestType.typeHash.toHexString(),
-    requestType.typeString,
-  ]);
-  store.remove("MetaTxRequestType", requestTypeId);
-}
-
-export function handleExecutionSuccess(event: ExecutionSuccessEvent): void {
-  logInvocation("handleExecutionSuccess", event);
-
-  let txHash = event.params.txHash;
-  let execution = new MetaTxExecution(txHash.toHexString());
-  execution.txHash = txHash;
-  log.info("creating MetaTxExecution hash [{}]", [txHash.toHexString()]);
-  execution.createdAt = eventUTCMillis(event);
-  execution.save();
 }
 
 export function handleMigrateLicensedUser(
